changed datatypes of floats to decimals

+++ SQL_SCHEMA_CREATION.sql
@@ -14,10 +14,10 @@ CREATE TABLE Country (
 
 CREATE TABLE EmissionData (
     EmissionID INT PRIMARY KEY,
-    CarbonDioxideEmissions FLOAT,
-    NitrousOxideEmissions FLOAT,
-    TotalGHGEmissions FLOAT,
-    MethaneEmissions FLOAT,
+    CarbonDioxideEmissions DECIMAL,
+    NitrousOxideEmissions DECIMAL,
+    TotalGHGEmissions DECIMAL,
+    MethaneEmissions DECIMAL,
     CountryID INT,
     FOREIGN KEY (CountryID) REFERENCES Country(CountryID)
 );
@@ -25,18 +25,18 @@ CREATE TABLE EmissionData (
 CREATE TABLE HealthSystem (
     HealthSystemID INT PRIMARY KEY,
     CountryID INT,
-    CompletenessOfBirthRegistration FLOAT,
-    CompletenessOfDeathRegistration FLOAT,
+    CompletenessOfBirthRegistration DECIMAL,
+    CompletenessOfDeathRegistration DECIMAL,
     FOREIGN KEY (CountryID) REFERENCES Country(CountryID)
 );
 
 CREATE TABLE HealthExpenditure (
     HealthSystemID INT PRIMARY KEY,
-    PercentOfGDP FLOAT,
-    PercentOfPublicFunded FLOAT,
-    PercentOutOfPocketCosts FLOAT,
-    PerCapita FLOAT,
-    ExternalHealthExpenditure FLOAT,
+    PercentOfGDP DECIMAL,
+    PercentOfPublicFunded DECIMAL,
+    PercentOutOfPocketCosts DECIMAL,
+    PerCapita DECIMAL,
+    ExternalHealthExpenditure DECIMAL,
     FOREIGN KEY (HealthSystemID) REFERENCES HealthSystem(HealthSystemID)
 );
 
@@ -44,7 +44,7 @@ CREATE TABLE HealthWorkers (
     HealthSystemID INT PRIMARY KEY,
     NumberOfNursesMidwives INT,
     NumberOfSpecialistSurgicalWorkforce INT,
-    NumberOfPhysicians FLOAT,
+    NumberOfPhysicians DECIMAL,
     FOREIGN KEY (HealthSystemID) REFERENCES HealthSystem(HealthSystemID)
 );
 
@@ -52,10 +52,10 @@ CREATE TABLE HealthRiskFactors (
     HealthRiskFactorID INT PRIMARY KEY,
     CountryID INT,
     HealthSystemID INT,
-    MaleTobaccoUse FLOAT,
-    FemaleTobaccoUse FLOAT,
-    IncidenceOfTuberculosis FLOAT,
-    PrevalenceOfDiabetes FLOAT,
+    MaleTobaccoUse DECIMAL,
+    FemaleTobaccoUse DECIMAL,
+    IncidenceOfTuberculosis DECIMAL,
+    PrevalenceOfDiabetes DECIMAL,
     FOREIGN KEY (CountryID) REFERENCES Country(CountryID),
     FOREIGN KEY (HealthSystemID) REFERENCES HealthSystem(HealthSystemID)
 );
@@ -63,17 +63,17 @@ CREATE TABLE HealthRiskFactors (
 CREATE TABLE PopulationCauseOfDeath (
     DeathID INT PRIMARY KEY,
     HealthRiskFactorID INT,
-    Injuries FLOAT,
-    NonCommunicableDiseases FLOAT,
-    CommunicableDiseases FLOAT,
+    Injuries DECIMAL,
+    NonCommunicableDiseases DECIMAL,
+    CommunicableDiseases DECIMAL,
     FOREIGN KEY (HealthRiskFactorID) REFERENCES HealthRiskFactors(HealthRiskFactorID)
 );
 
 CREATE TABLE PrevalenceOfHIV (
     IncidenceOfHIV INT PRIMARY KEY,
-    YouthMale FLOAT,
-    YouthFemale FLOAT,
-    WomanAge15Plus FLOAT
+    YouthMale DECIMAL,
+    YouthFemale DECIMAL,
+    WomanAge15Plus DECIMAL
 );
